requirements.txt
python-dotenv
websockets
redis
httpx
uvloop; platform_system!='Windows'
structlog
pydantic
pydantic-settings
boto3
